Custom logger class
